Move activity panel registration to its own folder
