--- client/header/fills.js
@@ -1,27 +1,0 @@
-/**
- * External dependencies
- */
-import { registerPlugin } from '@wordpress/plugins';
-
-/**
- * Internal dependencies
- */
-import ActivityPanel from '~/activity-panel';
-import { WooHeaderItem } from './utils';
-
-const ActivityPanelHeaderItem = () => (
-	<WooHeaderItem>
-		{ ( { isEmbedded, query } ) => {
-			if ( ! window.wcAdminFeatures[ 'activity-panels' ] ) {
-				return null;
-			}
-
-			return <ActivityPanel isEmbedded={ isEmbedded } query={ query } />;
-		} }
-	</WooHeaderItem>
-);
-
-registerPlugin( 'activity-panel-header-item', {
-	render: ActivityPanelHeaderItem,
-	scope: 'woocommerce-admin',
-} );
